Version 4.9.6-1
 * fixed version number in configure.ac
 * set library xx_VERSION_INFO to 5.0.0 for all libraries in metlibs
   -- need to update them when starting to compile as shared libraries
 * qUtilities/CoClient starts coserver in background; interface change
 * removed miString from qUtilities/miSliderWidget
 * removed some compiler and lintian warnings

+++ WdbQueries.cc
@@ -146,7 +146,7 @@ std::string CACHEQUERY(std::string model, std::string run, std::vector<std::stri
   else {
     query << "ARRAY[";
     bool first=true;
-    for( int i=0 ; i < parameters.size();i++) {
+    for(size_t i=0 ; i < parameters.size();i++) {
 
       if(boost::find_first(parameters[i],RAW())){
         hasRaw=true;
